Get cadquery and openscad working with more reuse on the backend
