+++ api/src/docker/common/entrypoint.sh
