+++ api/src/docker/openscad/runScad.js
@@ -1,6 +1,4 @@
-const { exec } = require('child_process')
-const { promises } = require('fs')
-const { writeFile } = promises
+const { makeFile, runCommand } = require('../common/utils')
 const { nanoid } = require('nanoid')
 
 module.exports.runScad = async ({
@@ -14,7 +12,7 @@ module.exports.runScad = async ({
     } = {},
   } = {}, // TODO add view settings
 } = {}) => {
-  const tempFile = await makeFile(file)
+  const tempFile = await makeFile(file, '.scad', nanoid)
   const { x: rx, y: ry, z: rz } = rotation
   const { x: px, y: py, z: pz } = position
   const cameraArg = `--camera=${px},${py},${pz},${rx},${ry},${rz},${dist}`
@@ -30,7 +28,7 @@ module.exports.runScad = async ({
 }
 
 module.exports.stlExport = async ({ file } = {}) => {
-  const tempFile = await makeFile(file)
+  const tempFile = await makeFile(file, '.scad', nanoid)
 
   try {
     const result = await runCommand(
@@ -41,37 +39,4 @@ module.exports.stlExport = async ({ file } = {}) => {
   } catch (error) {
     return { error, tempFile }
   }
-}
-
-async function makeFile(file) {
-  const tempFile = 'a' + nanoid() // 'a' ensure nothing funny happens if it start with a bad character like "-", maybe I should pick a safer id generator :shrug:
-  console.log(`file to write: ${file}`)
-
-  await runCommand(`mkdir /tmp/${tempFile}`)
-  await writeFile(`/tmp/${tempFile}/main.scad`, file)
-  return tempFile
-}
-
-async function runCommand(command, timeout = 5000) {
-  return new Promise((resolve, reject) => {
-    exec(command, (error, stdout, stderr) => {
-      if (error) {
-        console.log(`error: ${error.message}`)
-        console.log(`stderr: ${stderr}`)
-        console.log(`stdout: ${stdout}`)
-        reject(stdout || stderr) // it seems random if the message is in stdout or stderr, but not normally both
-        return
-      }
-      if (stderr) {
-        console.log(`stderr: ${stderr}`)
-        resolve(stderr)
-        return
-      }
-      console.log(`stdout: ${stdout}`)
-      resolve(stdout)
-    })
-    setTimeout(() => {
-      reject('timeout')
-    }, timeout)
-  })
 }
